[enh] Allow to install a given git reference of an app
This quickly adds support for git URL with /tree/ inside in app installation.
It is intended to ease tests of apps which use new manifest format and helpers
for the 2.4 release. An example usage:

 $ yunohost app install https://github.com/jeromelebleu/roundcube_ynh/tree/dev

As it will certainly change with the app category refactoring, it is not
documented and should be used with care.