Support multiple TXT entries for TLD
The dig of TXT for @ can returns multiple entries.
In that case, the DNS diagnosis fails.

The modification preserves the handling of DMARC and the likes
which use a single entry and a specfic domain name.

For single entry list, the behavior is preserved.

If mutliple TXT entries are defined for @, only the v=spf1
one is returned.

Signed-off-by: Nicolas Palix <nicolas.palix@imag.fr>