Add main domain into hosts file
For apps which send mail (or receive) which need a valid TLS certificate the way to make it work is to pass a valid domain (and not localhost), so the target domain match with the provided certificate.
But postfix and dovecot refuse app authentication from external IP. So we need to force the request on local interface (with the public domain).