Provision database when calling ynh_redis_get_free_db
The problem:
When we call redis_get_free_db, the user may assume that it reserves the
database returned. But in fact it does not until a key is created in
this DB.
The result is the following:
 ```bash
 # Let's assume redis is fresly installed and thus empty
 redis_db_1=$(ynh_redis_get_free_db) # returns 1
 redis_db_2=$(ynh_redis_get_free_db) # returns 1 too
 ```

The solution:
This commit ensures that a key is set in the database.
Furthermore, it is assigned to the name of the app, to facilitate the
debugging.