Forms: serialize None values as empty string.
Before this commit, None values were serialized into yaml as None, which
is considered as a string by yaml.
This led to errors when deserializing (for example, when the option is a
Number).

With this commit, None values are serialized as empty strings, which are
serialized as None by the BaseInputOption.