Ensure that ynh_string_random returns a result (#2181)
* test ynh_string_random

* Ensure that the ynh_string_random returns a result

this function can be passed a filter and a length.
But when combined, prior to this fix, we might not
get a result. For example, by running several times the following
command:
`ynh_string_random --length=64 --filter='a-f0-9'`

This is because we used to limit the number of characters output by
/dev/urandom

Now we read as many as random characters as needed to get a random
string given the passed filter and length.

Credits to Salamandar for the fix and to tituspijean for discovering the
issue.

Co-authored-by: Salamandar <6552989+Salamandar@users.noreply.github.com>

---------

Co-authored-by: Salamandar <6552989+Salamandar@users.noreply.github.com>